Bashrc: Remove -z and --append-verify from rsyncopts.
It seems that I missed a tiny little details in the man page about
append-verify: "If a file needs to be transferred and its size on the receiver
is the same or longer than the size on the sender, the file is skipped."

Also disable -z as it slows down local transfers more than speeds them
up. Will set -z manually on remote transfers is I need to.