I3status: Add last_full_capacity=true to config.
I3status battery monitor yields 76% upon fully loaded battery on one of
my laptops. Turns out it defaults to use the design capacity of the
battery as 100%. So, the older the battery becomes the lower the "fully
loaded" percentage becomes. This is a very realistic way of monitoring,
however, a bit painful.
Man page suggests to add last_full_capacity = true to config to fool
yourself.

I want to fool myself..